split client into smaller classes
