+++ lib/modules/platform/space/client/code-review.ts
@@ -1,0 +1,14 @@
+import {SpaceCodeReviewReadClient} from "./code-review-read";
+import {SpaceCodeReviewWriteClient} from "./code-review-write";
+import type {SpaceHttp} from "../../../../util/http/space";
+
+export class SpaceCodeReviewClient {
+  read: SpaceCodeReviewReadClient
+  write: SpaceCodeReviewWriteClient
+
+  constructor(http: SpaceHttp) {
+    this.read = new SpaceCodeReviewReadClient(http)
+    this.write = new SpaceCodeReviewWriteClient(http)
+  }
+
+}
